requirements.txt
# Web Framework
streamlit==1.31.0
flask==3.0.0

# Web Scraping
beautifulsoup4==4.12.3
newspaper3k==0.2.8
scrapy==2.11.0
requests==2.31.0
aiohttp==3.9.1
lxml==5.1.0

# Sentiment Analysis
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Database
sqlalchemy==2.0.25

# Visualization
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
